requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
catboost
xgboost
randomsearch
comet_ml
python-dotenv
-e .